multiple changes

+++ src/http/client/dl_head.c
@@ -10,7 +10,7 @@ int http_parse_headers_line (http_client_t * http, string_t * line) {
   string_t param;
   if (matchi_string_equal (line, "Content-Length: (%d+)", &param) > 0) {
     http->sz = atoi (param.ptr);
-    if (http->debug & DEBUG_HTTP_FILTER) printf ("  content length is %ld\n", http->sz);
+    if (http->debug & DEBUG_HTTP_FILTER) printf ("  content length is %lld\n", (long long)http->sz);
   } else if (matchi_string (line, "Transfer-Encoding:%s*") > 0) {
     if (matchi_string (line, "chunked") > 0) {
       if (http->debug & DEBUG_HTTP_FILTER) printf ("  transfer encoding is chunked\n");
